verkada-gcal/requirements.txt
requests
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
pytz
